split DocumentPreviewAPI to separate file document_preview.py
